Resolve some MS compiler warnings (in picky mode)
warning C4267: 'argument' : conversion from 'size_t' to 'DWORD', possible loss of data
warning C4800: 'BOOL' : forcing value to bool 'true' or 'false' (performance warning)
warning C4244: 'return' : conversion from 'double' to 'int', possible loss of data

+++ common/trace_file_snappy.cpp
@@ -397,7 +397,7 @@ bool SnappyFile::rawSkip(size_t length)
 
 int SnappyFile::rawPercentRead()
 {
-    return 100 * (double(m_stream.tellg()) / double(m_endPos));
+    return int(100 * (double(m_stream.tellg()) / double(m_endPos)));
 }
